Fix workspace reliability issues
