+++ apps/server/package.json
@@ -16,7 +16,7 @@
     "@OpenDiagram/harness": "workspace:*",
     "@ai-sdk/google": "^4.0.5",
     "@ai-sdk/groq": "^4.0.5",
-    "@ai-sdk/openai": "^2.0.0",
+    "@ai-sdk/openai": "^4.0.5",
     "ai": "catalog:",
     "better-auth": "catalog:",
     "dotenv": "catalog:",
