+++ apps/server/src/routes/github.ts
@@ -1,12 +1,13 @@
 import { auth } from "@OpenDiagram/auth";
-import { db } from "@OpenDiagram/db";
+import { and, db, eq } from "@OpenDiagram/db";
+import { githubImportJob } from "@OpenDiagram/db/schema/github-import-job";
 import { project } from "@OpenDiagram/db/schema/project";
 import { projectFile } from "@OpenDiagram/db/schema/project-file";
 import { z } from "zod";
 import { Hono } from "hono";
 import type { Context } from "hono";
 import { indexRepositoryMemory } from "../lib/project-memory";
-import { cloneAndBuildRepositoryDoc } from "../lib/repo-documentation";
+import { cleanupRepositoryClone, cloneAndBuildRepositoryDoc } from "../lib/repo-documentation";
 
 type GitHubRepository = {
   id: number;
@@ -45,6 +46,7 @@ export const githubRoute = new Hono();
 export const githubImportRoute = new Hono();
 
 const importJobs = new Map<string, GitHubImportJob>();
+const IMPORT_JOB_STALE_MS = 10 * 60 * 1000;
 
 githubRoute.get("/repositories", async (c) => {
   const token = await getGitHubAccessToken(c.req.raw.headers);
@@ -109,11 +111,11 @@ githubImportRoute.get("/github/:jobId", async (c) => {
   const session = await auth.api.getSession({ headers: c.req.raw.headers });
   if (!session) return c.json({ error: "Connect GitHub before importing repositories." }, 401);
 
-  const job = importJobs.get(c.req.param("jobId"));
+  const job = await getImportJob(c.req.param("jobId"), session.user.id);
   if (!job || job.userId !== session.user.id)
     return c.json({ error: "Import job not found." }, 404);
 
-  return c.json({ job: toPublicJob(job) });
+  return c.json({ job: toPublicJob(await failStaleImportJob(job)) });
 });
 
 async function importGitHubRepository(c: Context) {
@@ -167,7 +169,7 @@ async function importGitHubRepository(c: Context) {
     return c.json({ error: "Invalid response from GitHub." }, 502);
   }
 
-  const job = createImportJob({ repoFullName: repo.full_name, userId: authResult.userId });
+  const job = await createImportJob({ repoFullName: repo.full_name, userId: authResult.userId });
   void runImportJob({ jobId: job.id, repo, token: authResult.token });
 
   return c.json({ job: toPublicJob(job) }, 202);
@@ -200,7 +202,10 @@ async function getGitHubAuth(headers: Headers) {
   }
 }
 
-function createImportJob(input: { repoFullName: string; userId: string }): GitHubImportJob {
+async function createImportJob(input: {
+  repoFullName: string;
+  userId: string;
+}): Promise<GitHubImportJob> {
   const now = new Date().toISOString();
   const job: GitHubImportJob = {
     id: crypto.randomUUID(),
@@ -214,24 +219,30 @@ function createImportJob(input: { repoFullName: string; userId: string }): GitHu
     updatedAt: now,
   };
   importJobs.set(job.id, job);
+  await db.insert(githubImportJob).values(toJobRow(job));
   return job;
 }
 
 async function runImportJob(input: { jobId: string; repo: GitHubRepository; token: string }) {
   const job = importJobs.get(input.jobId);
   if (!job) return;
+  let repoPathToCleanup: string | null = null;
 
   try {
     const importedAt = new Date().toISOString();
-    updateImportJob(job.id, { status: "cloning", message: "Cloning repository to server" });
+    await updateImportJob(job.id, { status: "cloning", message: "Cloning repository to server" });
     const documentation = await cloneAndBuildRepositoryDoc({
       repoFullName: input.repo.full_name,
       defaultBranch: input.repo.default_branch,
       token: input.token,
       importedAt,
     });
+    repoPathToCleanup = documentation.repoPath;
 
-    updateImportJob(job.id, { status: "documenting", message: "Creating documentation file" });
+    await updateImportJob(job.id, {
+      status: "documenting",
+      message: "Creating documentation file",
+    });
     const [projectRow] = await db
       .insert(project)
       .values({
@@ -244,7 +255,6 @@ async function runImportJob(input: { jobId: string; repo: GitHubRepository; toke
           defaultBranch: input.repo.default_branch,
           htmlUrl: input.repo.html_url,
           importedAt,
-          repoPath: documentation.repoPath,
           commitSha: documentation.commitSha,
         },
       })
@@ -260,21 +270,20 @@ async function runImportJob(input: { jobId: string; repo: GitHubRepository; toke
       spec: documentation.provenance,
     });
 
-    updateImportJob(job.id, { status: "indexing", message: "Indexing repository memory" });
+    await updateImportJob(job.id, { status: "indexing", message: "Indexing repository memory" });
     const memoryResult = await indexRepositoryMemory({
       projectId: projectRow.id,
       userId: job.userId,
       repoFullName: input.repo.full_name,
       branch: input.repo.default_branch,
       commitSha: documentation.commitSha,
-      repoPath: documentation.repoPath,
       sourceDocuments: documentation.sourceDocuments,
     }).catch((error) => ({
       status: "failed",
       error: error instanceof Error ? error.message : "Repository memory indexing failed.",
     }));
 
-    updateImportJob(job.id, {
+    await updateImportJob(job.id, {
       status: "done",
       message:
         memoryResult?.status === "ready"
@@ -283,25 +292,102 @@ async function runImportJob(input: { jobId: string; repo: GitHubRepository; toke
       project: { id: projectRow.id, name: projectRow.name },
     });
   } catch (error) {
-    updateImportJob(job.id, {
+    await updateImportJob(job.id, {
       status: "failed",
       message: "Repository import failed",
       error: error instanceof Error ? error.message : "Repository import failed.",
     });
+  } finally {
+    if (repoPathToCleanup) {
+      await cleanupRepositoryClone(repoPathToCleanup).catch((error) => {
+        console.error("Failed to clean up repository clone:", error);
+      });
+    }
   }
 }
 
-function updateImportJob(
+async function updateImportJob(
   jobId: string,
   values: Partial<Pick<GitHubImportJob, "status" | "message" | "error" | "project">>,
 ) {
   const job = importJobs.get(jobId);
   if (!job) return;
 
-  importJobs.set(jobId, { ...job, ...values, updatedAt: new Date().toISOString() });
+  const nextJob = { ...job, ...values, updatedAt: new Date().toISOString() };
+  importJobs.set(jobId, nextJob);
+
+  await db
+    .update(githubImportJob)
+    .set({
+      status: nextJob.status,
+      message: nextJob.message,
+      error: nextJob.error,
+      projectId: nextJob.project?.id ?? null,
+      projectName: nextJob.project?.name ?? null,
+      updatedAt: nextJob.updatedAt,
+    })
+    .where(eq(githubImportJob.id, jobId));
 }
 
 function toPublicJob(job: GitHubImportJob) {
   const { userId: _userId, ...publicJob } = job;
   return publicJob;
+}
+
+async function getImportJob(jobId: string, userId: string) {
+  const cached = importJobs.get(jobId);
+  if (cached?.userId === userId) return cached;
+
+  const [row] = await db
+    .select()
+    .from(githubImportJob)
+    .where(and(eq(githubImportJob.id, jobId), eq(githubImportJob.userId, userId)));
+
+  if (!row) return null;
+
+  const job = fromJobRow(row);
+  importJobs.set(job.id, job);
+  return job;
+}
+
+async function failStaleImportJob(job: GitHubImportJob) {
+  if (job.status === "done" || job.status === "failed") return job;
+  if (Date.now() - Date.parse(job.updatedAt) < IMPORT_JOB_STALE_MS) return job;
+
+  await updateImportJob(job.id, {
+    status: "failed",
+    message: "Repository import interrupted",
+    error: "Repository import was interrupted. Please start the import again.",
+  });
+
+  return importJobs.get(job.id) ?? job;
+}
+
+function toJobRow(job: GitHubImportJob): typeof githubImportJob.$inferInsert {
+  return {
+    id: job.id,
+    userId: job.userId,
+    repoFullName: job.repoFullName,
+    status: job.status,
+    message: job.message,
+    error: job.error,
+    projectId: job.project?.id ?? null,
+    projectName: job.project?.name ?? null,
+    createdAt: job.createdAt,
+    updatedAt: job.updatedAt,
+  };
+}
+
+function fromJobRow(row: typeof githubImportJob.$inferSelect): GitHubImportJob {
+  return {
+    id: row.id,
+    userId: row.userId,
+    repoFullName: row.repoFullName,
+    status: row.status,
+    message: row.message,
+    error: row.error,
+    project: row.projectId && row.projectName ? { id: row.projectId, name: row.projectName } : null,
+    createdAt: row.createdAt,
+    updatedAt: row.updatedAt,
+  };
 }
