+++ packages/db/src/schema/index.ts
@@ -1,3 +1,4 @@
 export * from "./auth";
+export * from "./github-import-job";
 export * from "./project";
 export * from "./project-file";
